🌱 initial commit
Add first version of aws-mfa function.

Ask device ARN at the first usage, then create a file to store the
device ARN.
My first draft of the function got a token code as an argument, but
changed to ask when the function triggered. There are couple of reasons
I prefer this:

- Don't need to type token code with function (easy to call from history)
- Hide token code from history

The second one doen't really matter, but just nice :)